public.json: Factor 'updateCategory' out from 'category'
This gives us space to document 'ancestors' and 'depth' without
implying that you can upload them.  Also drop 'id' from
updateCategory, since that's either generated by the backend (for
POST) or carried externally (for PUT, where it's in the path).

Also convert the descriptions to sentences to avoid adding new
updateCategory descriptions that aren't sentences or which diverge
from the category descriptions.