public.json: Add ?inline=person to /addresses
So folks can go from a given address back to the associated person
(e.g. "which customer is +1 800 555 0100?" could be
GET /address?telephone=%2b1%20800%20555%200100&inline=person.email).
For example, this allows customer lookups by phone number for customer
service reps fielding incoming SMS messages and/or voice calls.

While I have (very) long-term plans for consolidating addresses so
that several drops can share the same addresses, I doubt we'll have
multiple customers sharing the same address, even if the parameters
are identical.  If that happens, we'll have to replace the 'person'
inline with 'people', but we can cross that bridge if/when we come to
it.

I've also extended the address ?inline=... support to the other
address endpoints for symmetry.