public.json: Add order.rewards-allocation
So folks can stage pending reward debits (e.g. "use up to 12.34
rewards on this order").

In a meeting with Casey, David, and Tom, we considered adding
reward-credit and reward-debit properties with IDs for those reward
entries with ?inline=reward-debit, etc. to access them.  We decided to
add this order property instead, primarily because you can set it on
open orders.  Customers will likely set this value as they prepare
their order for checkout, so it will intially be set for an open
order.  We want to save that open-order preference to the database, to
support customers who get part way through checkout and then wish to
complete the process from another device.  We don't want reward
allocations for open orders locking otherwise-available rewards.

With this approach, users have a single endpoint to hit for the atomic
"place this order" operation, which transitions the order status and
locks the allocated rewards (among other things ;).  And they can set
the allocation ahead of time or not, as they wish, for their own
benefit, without worrying about open-order allocations having teeth.
Users interested in the reward entries will have to make a separate
/rewards?order=12 request, but we're ok with that.