public.json: Add _price.rewards and friends
So users can see what they'll earn for a given packaging before they
add an order line.  Also add 'minimum' entries to these floats to show
they are >= 0.

The *-standard forms are a machine-readable analog to the dollar
'discount' property, and the frontend can use them to calculate "bonus
rewards" UIs and similar.

In a meeting with Casey, David, and Tom and subsequent discussion in
[1], we considered setting the *-standard properties in all cases, or
only when that value was lower than the actual value.

We could not think of a desirable use case for the standard
information when it exceeded the actual rewards for the packaging, and
by only setting rewards-standard when < rewards we would have saved a
few bytes and avoided exposing unusable information to the API callers
(effectively collapsing the < rewards and = rewards cases together).
However, consuming JavaScript for bonus rewards would either be the
less obvious:

  if (price['rewards-standard'] !== undefined)

or the redundant:

  if (price['rewards-standard'] !== undefined && price.rewards > price['rewards-standard'])

By always setting the standard properties (as in this commit), we
allow for the more obvious:

  if (price.rewards > price['rewards-standard'])

without redundancy or the need to check for the existence of
rewards-standard in the response.

We aren't sure we need rewards-per-unit(-standard), and might remove
them in the future if it turns out we don't.

[1]: https://github.com/azurestandard/api-spec/pull/230#discussion_r191887254